Move notifications debug section to Debugging screen
The notifications debug controls (rate limits, local push notifications,
push ID sharing, and reset push ID) previously lived in a "Debug" section
at the bottom of Settings > Notifications. Move them under
Settings > Debugging > Notifications so all debug-oriented tooling is
grouped together.

- Add NotificationDebugView with the moved controls, reachable via a new
  "Notifications" row in DebugView's top section.
- Extract push ID / reset logic into NotificationDebugViewModel and the
  share sheet into its own NotificationsShareSheet file.
- Strip the now-unused debug section, state, and view-model members from
  NotificationSettingsView.
