+++ Sources/App/Settings/Notifications/NotificationsShareSheet.swift
@@ -1,0 +1,11 @@
+import SwiftUI
+
+struct NotificationsShareSheet: UIViewControllerRepresentable {
+    let activityItems: [Any]
+
+    func makeUIViewController(context: Context) -> UIActivityViewController {
+        UIActivityViewController(activityItems: activityItems, applicationActivities: nil)
+    }
+
+    func updateUIViewController(_ controller: UIActivityViewController, context: Context) {}
+}
